Provide corresponding tests of the RNN model and training
